# openapi.yaml
openapi: 3.0.3
info:
  title: hterm REST API
  description: |
    HTTP/REST API for hterm - A web-based terminal server with MCP (Model Context Protocol) support.

    ## Authentication

    All endpoints (except `/api/config` and static assets) support two authentication modes:

    1. **HTTP Basic Auth**: Set via `--credential username:password` flag
       - Header: `Authorization: Basic <base64(username:password)>`

    2. **Reverse Proxy Auth**: Set via `--auth-header X-Remote-User` flag
       - Header: `X-Remote-User: <username>` (or custom header name)

    If neither is configured, authentication is disabled.

    ## Base Path

    All routes can be prefixed with a custom base path using the `--base-path` flag.
    Example: `--base-path /terminal` results in routes like `/terminal/api/config`

  version: 0.1.0
  contact:
    name: GitHub Repository
    url: https://github.com/Rishang/hterm
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:7681
    description: Default local server
  - url: https://localhost:7681
    description: Local server with TLS
...
paths:
...
  /api/tools/call:
    post:
      summary: Execute an MCP tool
      description: |
        Call a tool by name with arguments. Returns tool output in MCP format.

        **Available tools**: run_command, read_file, write_file, list_processes,
        list_files, list_tree, count_file_lines, read_file_size
      tags:
        - tools
      security:
        - BasicAuth: []
        - ProxyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCallRequest'
            examples:
              run_command:
                summary: Execute shell command
                value:
                  name: run_command
                  arguments:
                    command: "whoami"
                    timeout_secs: 30
              read_file:
                summary: Read file
                value:
                  name: read_file
                  arguments:
                    path: "/etc/hostname"
              write_file:
                summary: Write file
                value:
                  name: write_file
                  arguments:
                    path: "/tmp/test.txt"
                    content: "Hello, World!"
              list_files:
                summary: List directory
                value:
                  name: list_files
                  arguments:
                    path: "/home"
      responses:
        '200':
          description: Tool executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCallResponse'
              examples:
                success:
                  summary: Successful execution
                  value:
                    content:
                      - type: text
                        text: "root\n"
                    isError: false
                error:
                  summary: Command failed
                  value:
                    content:
                      - type: text
                        text: "ls: cannot access '/nonexistent': No such file or directory"
                    isError: true
        '400':
          description: Invalid tool name or arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /mcp/sse:
    get:
      summary: MCP SSE endpoint
      description: |
        Server-Sent Events endpoint for MCP protocol communication.

        Returns an SSE stream that includes:
        1. `endpoint` event with POST URL for messages
        2. `message` events with JSON-RPC 2.0 responses
      tags:
        - mcp
      security:
        - BasicAuth: []
        - ProxyAuth: []
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: endpoint
                data: /mcp/message?sessionId=550e8400-e29b-41d4-a716-446655440000

                event: message
                data: {"jsonrpc":"2.0","id":1,"result":{"protocolVersion":"2024-11-05"}}
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ToolCallRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Tool name to execute
          example: "run_command"
        arguments:
          type: object
          description: Tool-specific arguments
          additionalProperties: true
          example:
            command: "uptime"
    ToolCallResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [text]
              text:
                type: string
        isError:
          type: boolean
          description: Whether the tool execution resulted in an error
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        text/plain:
          schema:
            type: string
            example: "Unauthorized"
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication (configured with --credential)

    ProxyAuth:
      type: apiKey
      in: header
      name: X-Remote-User
      description: |
        Reverse proxy authentication header (configured with --auth-header).
        Header name is customizable via CLI flag.